pyproject.toml
[tool.poetry]
name = "render2d"
version = "0.1.0"
description = ""
authors = ["David Berthiaume <odlantern@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.22.3"
Pillow = "^9.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
